requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
azureopenai==0.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
dashscope==1.22.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.7.0
email-validator==2.1.0.post1
et_xmlfile==2.0.0
fasttext-wheel==0.9.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.53
langsmith==0.1.147
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
msal==1.31.1
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.61.1
openpyxl==3.1.2
orjson==3.10.15
packaging==23.2
pandas==2.2.1
pillow==10.2.0
propcache==0.2.1
psycopg2-binary==2.9.9
pybind11==2.13.6
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
pymongo==4.6.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-imap==1.0.0
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
XlsxWriter==3.2.2
yarl==1.18.3
